+++ Letterbook.IntegrationTests.sln
@@ -22,6 +22,10 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Development", "Development"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Letterbook.Core", "Letterbook.Core\Letterbook.Core.csproj", "{808895B8-D5F6-4DE3-AFF6-91C19E75B78C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Letterbook.Api", "Letterbook.Api\Letterbook.Api.csproj", "{E5A35B97-C660-48D5-B241-9C97CD1A5388}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Letterbook.Api.IntegrationTests", "Letterbook.Api.IntegrationTests\Letterbook.Api.IntegrationTests.csproj", "{AE1F64BB-A42E-4B22-A47E-4EB06A4CA59A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -47,5 +51,13 @@ Global
 		{808895B8-D5F6-4DE3-AFF6-91C19E75B78C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{808895B8-D5F6-4DE3-AFF6-91C19E75B78C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{808895B8-D5F6-4DE3-AFF6-91C19E75B78C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E5A35B97-C660-48D5-B241-9C97CD1A5388}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E5A35B97-C660-48D5-B241-9C97CD1A5388}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E5A35B97-C660-48D5-B241-9C97CD1A5388}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E5A35B97-C660-48D5-B241-9C97CD1A5388}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE1F64BB-A42E-4B22-A47E-4EB06A4CA59A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE1F64BB-A42E-4B22-A47E-4EB06A4CA59A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE1F64BB-A42E-4B22-A47E-4EB06A4CA59A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE1F64BB-A42E-4B22-A47E-4EB06A4CA59A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
